Enhancement: Project Structure for Improved Readability and Maintenance #15: Changed directory structure.
