+++ src/core/batch_tour_length_minimizer.py
@@ -3,7 +3,7 @@ import random
 import time
 import uuid
 
-from src.batch_handler.batch_tour_length_calculator import calculate_tour_length_s_shape_routing
+from src.core.batch_tour_length_calculator import calculate_tour_length_s_shape_routing
 
 def create_start_batches(orders, max_batch_size):
     """
