+++ src/core/join_item_information.py
